move:(component file moved from components to the admin folder)
