+++ frontend/src/App.tsx
@@ -6,7 +6,7 @@ import { Login } from './pages/Login';
 import { Register } from './pages/Register';
 import { Dashboard } from './pages/Dashboard';
 import { Profile } from './pages/Profile';
-import { ProtectedRoute } from './components/ProtectedRoute';
+import { ProtectedRoute } from './components/admin/ProtectedRoute';
 
 function App() {
   return (
