+++ frontend/src/components/admin/chart-area-interactive.tsx
