+++ frontend/src/components/admin/section-cards.tsx
