+++ frontend/src/pages/Dashboard.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { useAuth } from '../contexts/AuthContext';
 import { User, Calendar, Mail } from 'lucide-react';
 
-import { AppSidebar } from "@/components/app-sidebar"
-import { ChartAreaInteractive } from "@/components/chart-area-interactive"
-import { DataTable } from "@/components/data-table"
-import { SectionCards } from "@/components/section-cards"
-import { SiteHeader } from "@/components/site-header"
+import { AppSidebar } from "@/components/admin/app-sidebar"
+import { ChartAreaInteractive } from "@/components/admin/chart-area-interactive"
+import { DataTable } from "@/components/admin/data-table"
+import { SectionCards } from "@/components/admin/section-cards"
+import { SiteHeader } from "@/components/admin/site-header"
 import {
   SidebarInset,
   SidebarProvider,
